+++ src/utils/index.ts
@@ -5,3 +5,4 @@ export * from "./errors";
 export * from "./funding";
 export * from "./liquidity";
 export * from "./rpc";
+export * from "./slippage";